rename test folder to tests
