+++ tests/data/demo_import_functions.yml
@@ -1,7 +1,7 @@
 - config:
     name: "create user testsets."
     import_module_functions:
-        - test.data.custom_functions
+        - tests.data.custom_functions
     variable_binds:
         - TOKEN: debugtalk
         - json: {}
