+++ tests/data/simple_demo_auth_hardcode.yml
