+++ tests/test_apiserver_v2.py
@@ -1,24 +1,34 @@
-import requests
 import random
-from test.base import ApiServerUnittest
+import requests
 
-class TestApiServer(ApiServerUnittest):
+from tests.base import ApiServerUnittest
+
+
+class TestApiServerV2(ApiServerUnittest):
+
+    authentication = True
+
     def setUp(self):
-        super(TestApiServer, self).setUp()
+        super(TestApiServerV2, self).setUp()
         self.host = "http://127.0.0.1:5000"
         self.api_client = requests.Session()
         self.clear_users()
 
     def tearDown(self):
-        super(TestApiServer, self).tearDown()
+        super(TestApiServerV2, self).tearDown()
+
+    def test_index(self):
+        headers = self.prepare_headers()
+        resp = self.api_client.get(self.host, headers=headers)
+        self.assertEqual(200, resp.status_code)
 
     def clear_users(self):
         url = "%s/api/users" % self.host
-        return self.api_client.delete(url)
+        return self.api_client.delete(url, headers=self.prepare_headers())
 
     def get_users(self):
         url = "%s/api/users" % self.host
-        return self.api_client.get(url)
+        return self.api_client.get(url, headers=self.prepare_headers())
 
     def create_user(self, uid, name, password):
         url = "%s/api/users/%d" % (self.host, uid)
@@ -26,11 +36,12 @@ class TestApiServer(ApiServerUnittest):
             'name': name,
             'password': password
         }
-        return self.api_client.post(url, json=data)
+        headers = self.prepare_headers(data)
+        return self.api_client.post(url, headers=headers, json=data)
 
     def get_user(self, uid):
         url = "%s/api/users/%d" % (self.host, uid)
-        return self.api_client.get(url)
+        return self.api_client.get(url, headers=self.prepare_headers())
 
     def update_user(self, uid, name, password):
         url = "%s/api/users/%d" % (self.host, uid)
@@ -38,11 +49,12 @@ class TestApiServer(ApiServerUnittest):
             'name': name,
             'password': password
         }
-        return self.api_client.put(url, json=data)
+        headers = self.prepare_headers(data)
+        return self.api_client.put(url, headers=headers, json=data)
 
     def delete_user(self, uid):
         url = "%s/api/users/%d" % (self.host, uid)
-        return self.api_client.delete(url)
+        return self.api_client.delete(url, headers=self.prepare_headers())
 
     def test_clear_users(self):
         resp = self.clear_users()
@@ -114,7 +126,11 @@ class TestApiServer(ApiServerUnittest):
         expected_response = {
             'status_code': status_code,
         }
-        resp = self.api_client.post(url, json=expected_response)
+        resp = self.api_client.post(
+            url,
+            headers=self.prepare_headers(expected_response),
+            json=expected_response
+        )
         self.assertEqual(status_code, resp.status_code)
 
     def test_get_customized_response_headers(self):
@@ -125,13 +141,18 @@ class TestApiServer(ApiServerUnittest):
             }
         }
         url = "%s/customize-response" % self.host
-        resp = self.api_client.post(url, json=expected_response)
+        resp = self.api_client.post(
+            url,
+            headers=self.prepare_headers(expected_response),
+            json=expected_response
+        )
         self.assertIn('abc', resp.headers)
         self.assertIn('123', resp.headers['abc'])
 
     def test_get_token(self):
         url = "%s/api/token" % self.host
-        resp = self.api_client.get(url)
+        headers = self.prepare_headers()
+        resp = self.api_client.get(url, headers=headers)
         resp_json = resp.json()
         self.assertTrue(resp_json["success"])
         self.assertEqual(len(resp_json["token"]), 8)
