+++ tests/test_runner_v2.py
@@ -1,7 +1,7 @@
 import os
 import requests
 from ate import runner, exception, utils
-from test.base import ApiServerUnittest
+from tests.base import ApiServerUnittest
 
 class TestRunnerV2(ApiServerUnittest):
 
@@ -17,7 +17,7 @@ class TestRunnerV2(ApiServerUnittest):
 
     def test_run_single_testcase_yaml(self):
         testcase_file_path = os.path.join(
-            os.getcwd(), 'test/data/simple_demo_auth_hardcode.yml')
+            os.getcwd(), 'tests/data/simple_demo_auth_hardcode.yml')
         testcases = utils.load_testcases(testcase_file_path)
         testcase = testcases[0]["test"]
         success, _ = self.test_runner.run_test(testcase)
@@ -25,7 +25,7 @@ class TestRunnerV2(ApiServerUnittest):
 
     def test_run_single_testcase_json(self):
         testcase_file_path = os.path.join(
-            os.getcwd(), 'test/data/simple_demo_auth_hardcode.json')
+            os.getcwd(), 'tests/data/simple_demo_auth_hardcode.json')
         testcases = utils.load_testcases(testcase_file_path)
         testcase = testcases[0]["test"]
         success, _ = self.test_runner.run_test(testcase)
@@ -33,7 +33,7 @@ class TestRunnerV2(ApiServerUnittest):
 
     def test_run_testset_auth_yaml(self):
         testcase_file_path = os.path.join(
-            os.getcwd(), 'test/data/simple_demo_auth_hardcode.yml')
+            os.getcwd(), 'tests/data/simple_demo_auth_hardcode.yml')
         testsets = utils.load_testcases_by_path(testcase_file_path)
         results = self.test_runner.run_testset(testsets[0])
         self.assertEqual(len(results), 2)
@@ -41,7 +41,7 @@ class TestRunnerV2(ApiServerUnittest):
 
     def test_run_testsets_auth_yaml(self):
         testcase_file_path = os.path.join(
-            os.getcwd(), 'test/data/simple_demo_auth_hardcode.yml')
+            os.getcwd(), 'tests/data/simple_demo_auth_hardcode.yml')
         testsets = utils.load_testcases_by_path(testcase_file_path)
         results = self.test_runner.run_testsets(testsets)
         self.assertEqual(len(results), 1)
@@ -49,7 +49,7 @@ class TestRunnerV2(ApiServerUnittest):
 
     def test_run_testset_auth_json(self):
         testcase_file_path = os.path.join(
-            os.getcwd(), 'test/data/simple_demo_auth_hardcode.json')
+            os.getcwd(), 'tests/data/simple_demo_auth_hardcode.json')
         testsets = utils.load_testcases_by_path(testcase_file_path)
         results = self.test_runner.run_testset(testsets[0])
         self.assertEqual(len(results), 2)
@@ -57,7 +57,7 @@ class TestRunnerV2(ApiServerUnittest):
 
     def test_run_testsets_auth_json(self):
         testcase_file_path = os.path.join(
-            os.getcwd(), 'test/data/simple_demo_auth_hardcode.json')
+            os.getcwd(), 'tests/data/simple_demo_auth_hardcode.json')
         testsets = utils.load_testcases_by_path(testcase_file_path)
         results = self.test_runner.run_testsets(testsets)
         self.assertEqual(len(results), 1)
@@ -65,7 +65,7 @@ class TestRunnerV2(ApiServerUnittest):
 
     def test_run_testcase_template_yaml(self):
         testcase_file_path = os.path.join(
-            os.getcwd(), 'test/data/demo_template_separate.yml')
+            os.getcwd(), 'tests/data/demo_template_separate.yml')
         testcases = utils.load_testcases(testcase_file_path)
         success, _ = self.test_runner.run_test(testcases[0]["test"])
         self.assertTrue(success)
@@ -74,7 +74,7 @@ class TestRunnerV2(ApiServerUnittest):
 
     def test_run_testset_template_yaml(self):
         testcase_file_path = os.path.join(
-            os.getcwd(), 'test/data/demo_template_sets.yml')
+            os.getcwd(), 'tests/data/demo_template_sets.yml')
         testsets = utils.load_testcases_by_path(testcase_file_path)
         results = self.test_runner.run_testset(testsets[0])
         self.assertEqual(len(results), 2)
@@ -82,7 +82,7 @@ class TestRunnerV2(ApiServerUnittest):
 
     def test_run_testsets_template_yaml(self):
         testcase_file_path = os.path.join(
-            os.getcwd(), 'test/data/demo_template_sets.yml')
+            os.getcwd(), 'tests/data/demo_template_sets.yml')
         testsets = utils.load_testcases_by_path(testcase_file_path)
         results = self.test_runner.run_testsets(testsets)
         self.assertEqual(len(results), 1)
@@ -90,7 +90,7 @@ class TestRunnerV2(ApiServerUnittest):
 
     def test_run_testset_template_import_functions(self):
         testcase_file_path = os.path.join(
-            os.getcwd(), 'test/data/demo_import_functions.yml')
+            os.getcwd(), 'tests/data/demo_import_functions.yml')
         testsets = utils.load_testcases_by_path(testcase_file_path)
         results = self.test_runner.run_testset(testsets[0])
         self.assertEqual(len(results), 2)
@@ -98,7 +98,7 @@ class TestRunnerV2(ApiServerUnittest):
 
     def test_run_testsets_template_import_functions(self):
         testcase_file_path = os.path.join(
-            os.getcwd(), 'test/data/demo_import_functions.yml')
+            os.getcwd(), 'tests/data/demo_import_functions.yml')
         testsets = utils.load_testcases_by_path(testcase_file_path)
         results = self.test_runner.run_testsets(testsets)
         self.assertEqual(len(results), 1)
