+++ tests/test_task.py
@@ -1,7 +1,7 @@
 import os
 import random
 import requests
-from test.base import ApiServerUnittest
+from tests.base import ApiServerUnittest
 from ate import task, utils
 
 class TestMain(ApiServerUnittest):
@@ -14,7 +14,7 @@ class TestMain(ApiServerUnittest):
         return requests.delete(url)
 
     def test_create_suite(self):
-        testcase_file_path = os.path.join(os.getcwd(), 'test/data/simple_demo_no_auth.yml')
+        testcase_file_path = os.path.join(os.getcwd(), 'tests/data/simple_demo_no_auth.yml')
         testsets = utils.load_testcases_by_path(testcase_file_path)
         suite = task.create_suite(testsets[0])
         self.assertEqual(suite.countTestCases(), 2)
@@ -22,7 +22,7 @@ class TestMain(ApiServerUnittest):
             self.assertIsInstance(testcase, task.ApiTestCase)
 
     def test_create_task(self):
-        testcase_file_path = os.path.join(os.getcwd(), 'test/data/simple_demo_no_auth.yml')
+        testcase_file_path = os.path.join(os.getcwd(), 'tests/data/simple_demo_no_auth.yml')
         task_suite = task.create_task(testcase_file_path)
         self.assertEqual(task_suite.countTestCases(), 2)
         for suite in task_suite:
